requirements.txt
# Core Scientific Computing
numpy==2.3.4
scipy==1.16.3
scikit-learn==1.7.2
statsmodels==0.14.5
pandas==2.3.3
matplotlib==3.8.0
seaborn==0.13.2

# PyTorch and Deep Learning
torch==2.9.0
torchvision==0.24.0
torchaudio==2.8.0
triton==3.5.0

# NVIDIA CUDA Libraries
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90

# Transformers and Model Training
transformers==4.57.1
accelerate==1.11.0
bitsandbytes==0.48.2
peft==0.17.1
trl==0.25.0
datasets==4.4.1
tokenizers==0.22.1
sentencepiece==0.1.99
safetensors==0.6.2

# HuggingFace Hub
huggingface-hub==0.36.0
hf-xet==1.2.0

# LangChain
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-classic==1.0.0
langchain-community==0.3.31
langchain-core==0.3.79
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38

# API Clients
openai==2.6.1
anthropic==0.72.0
wandb==0.22.3

# NLP and Evaluation
rouge-score==0.1.2
tiktoken==0.12.0
nltk==3.9.2
wikipedia==1.4.0

# Vector Database
chromadb==1.3.0

# Utilities
fire==0.7.1
python-dotenv==1.0.0
black==25.9.0
tqdm==4.67.1
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pydantic-extra-types==2.10.6
click==8.1.7
PyYAML==6.0.1
typing-extensions==4.15.0
packaging==23.2
filelock==3.20.0
fsspec==2023.10.0
requests==2.32.5
certifi==2025.10.5

# Data Processing
pyarrow==22.0.0
pyarrow-hotfix==0.6
dill==0.3.7
multiprocess==0.70.15
xxhash==3.6.0

# Async and Web
aiohttp==3.9.1
aiosignal==1.3.1
aioprometheus==23.12.0
anyio==3.7.1
httpx==0.28.1
httpcore==1.0.2
httpx-sse==0.4.3
httptools==0.7.1
h11==0.14.0
websocket-client==1.9.0
websockets==15.0.1

# FastAPI and Server
fastapi==0.120.2
fastapi-cli==0.0.14
starlette==0.49.1
uvicorn==0.38.0
uvloop==0.22.1
python-multipart==0.0.20
watchfiles==1.1.1

# Monitoring and Logging
coloredlogs==15.0.1
python-json-logger==4.0.0
rich==14.2.0
sentry-sdk==2.43.0
posthog==5.4.0

# Serialization and Data Formats
msgpack==1.1.2
msgspec==0.19.0
orjson==3.11.4
ormsgpack==1.11.0
protobuf==3.20.3
flatbuffers==25.9.23
cbor2==5.7.1

# JSON and Schema
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonpatch==1.33
jsonpointer==3.0.0
partial-json-parser==0.2.1.1.post6

# Compression
Brotli==1.0.9
zstandard==0.25.0

# Cryptography and Security
bcrypt==5.0.0
cffi==2.0.0
pycparser==2.23

# System and OS
psutil==5.9.0
setproctitle==1.3.7

# String and Text Processing
charset-normalizer==3.3.2
regex==2023.10.3
Jinja2==3.1.6
MarkupSafe==2.1.1
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2

# Math and Scientific
sympy==1.14.0
mpmath==1.3.0
numba==0.61.2
llvmlite==0.44.0

# Machine Learning Support
joblib==1.3.2
threadpoolctl==3.6.0
cloudpickle==3.1.1

# Plotting and Visualization
contourpy==1.2.0
cycler==0.11.0
fonttools==4.25.0
kiwisolver==1.4.4
pillow==12.0.0
pyparsing==3.0.9
adjustText==0.7.3

# Date and Time
python-dateutil==2.8.2
pytz==2023.3.post1
tzdata==2023.3

# Networking and URLs
urllib3==2.1.0
idna==3.6
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
oauthlib==3.3.1

# Parsing and Grammars
lark==1.2.2
interegular==0.3.3
ply==3.11

# Distributed Computing
ray==2.51.0

# Google Cloud and Auth
google-auth==2.42.0
googleapis-common-protos==1.71.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
rsa==4.9.1

# Monitoring and Metrics
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.13
grpcio==1.76.0

# Audio and Media
soundfile==0.13.1
soxr==1.0.0
opencv-python-headless==4.11.0.86

# Git Integration
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# Database
SQLAlchemy==2.0.44
diskcache==5.6.3
PyPika==0.48.9

# Web Scraping
beautifulsoup4==4.14.2
soupsieve==2.8

# Kubernetes
kubernetes==34.1.0

# Type Checking and Validation
annotated-types==0.7.0
typing-inspect==0.9.0
typing-inspection==0.4.2
types-pytz==2022.4.0.0
mypy_extensions==1.1.0

# CLI Tools
typer==0.20.0
shellingham==1.5.4
termcolor==3.2.0
colorama==0.4.6

# Caching and Hashing
cachetools==6.2.1
mmh3==5.2.0
blake3==1.0.8

# Data Classes and Serialization
dataclasses-json==0.6.7
marshmallow==3.26.1

# Retry and Backoff
backoff==2.2.1
tenacity==8.5.0

# Misc Dependencies
attrs==23.1.0
Deprecated==1.3.0
frozendict==2.4.6
frozenlist==1.4.0
greenlet==3.2.4
humanfriendly==10.0
importlib-metadata==6.8.0
importlib-resources==6.5.2
jiter==0.11.1
multidict==6.0.4
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.1
overrides==7.7.0
pathspec==0.12.1
patsy==1.0.2
platformdirs==4.5.0
referencing==0.37.0
rpds-py==0.28.0
six==1.16.0
sniffio==1.3.0
wrapt==2.0.0
yarl==1.9.3
zipp==3.17.0

# vLLM and Inference
vllm==0.11.0
xformers==0.0.32.post1
compressed-tensors==0.11.0
outlines==0.1.11
outlines-core==0.2.11
llguidance==0.7.30
xgrammar==0.1.25
lm-format-enforcer==0.11.3
gguf==0.17.1

# Specialized ML Tools
einops==0.8.1
ninja==1.13.0
onnxruntime==1.23.2
cupy-cuda12x==13.6.0
fastrlock==0.8.3
gmpy2==2.1.2
sacremoses==0.0.53
mistral-common==1.8.5
pytokens==0.2.0
quantile-python==1.1

# Build Tools
build==1.3.0
pyproject-hooks==1.2.2

# Additional Utilities
absl-py==2.3.1
airportsdata==20250909
annotated-doc==0.0.3
astor==0.8.1
depyf==0.19.0
dnspython==2.8.0
docstring-parser==0.17.0
durationpy==0.10
email-validator==2.3.0
fastapi-cloud-cli==0.3.1
openai-harmony==0.0.4
pandas-stubs==1.5.3.230203
procoder==0.0.1
py-cpuinfo==9.0.0
pybase64==1.4.2
pycountry==24.6.1
pyzmq==27.1.0
rich-toolkit==0.15.1
rignore==0.7.2
roman==5.1
tornado==6.3.3

# Note: toolemu is NOT listed here because it's a local modified package.
# It gets installed via `pip install -e .` (see setup.py)
